icl-task-diversity/requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
black==23.3.0
cached-property==1.5.2
cachetools==5.3.0
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
cloudpickle==2.2.1
clu==0.0.8
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.1.1
flatbuffers==23.3.3
flax==0.6.8
fonttools==4.39.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
isort==5.12.0
jax==0.4.8
jaxlib==0.4.7
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
-f https://storage.googleapis.com/jax-releases/libtpu_releases.html
libtpu-nightly==0.1.dev20230327
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.0.4
msgpack==1.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.7
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pathtools==0.1.2
Pillow==9.4.0
platformdirs==3.2.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.19.0
tensorstore==0.1.35
termcolor==2.2.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
hydra-core==1.3.2
